+++ common/src/main/java/com/github/teamfossilsarcheology/fossil/util/ModConstants.java
@@ -6,4 +6,5 @@ public class ModConstants {
     public static final String CREATE = "create";
     public static final String FARMERS = "farmersdelight";
     public static final String LITHOSTITCHED = "lithostitched";
+    public static final String MIDNIGHT_LIB = "midnightlib";
 }