+++ fabric/src/main/java/com/github/teamfossilsarcheology/fossil/fabric/mixin/MidnightConfigModernMixin.java
@@ -5,6 +5,7 @@ import com.github.teamfossilsarcheology.fossil.config.fabric.FabricConfigFix;
 import com.github.teamfossilsarcheology.fossil.config.fabric.FossilConfigImpl;
 import eu.midnightdust.lib.config.MidnightConfig;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
@@ -12,14 +13,17 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import java.lang.reflect.Field;
 
 /**
- * Mixin for MidnightLib with version <1.9.0
+ * Mixin for MidnightLib with version >=1.9.0
  */
-@Mixin(MidnightConfig.class)
-public class MidnightConfigMixin {
+@Mixin(targets = "eu.midnightdust.lib.config.MidnightConfig")
+public class MidnightConfigModernMixin {
 
-    @Inject(method = "init", remap = false, at = @At(value = "INVOKE", target = "Ljava/util/List;iterator()Ljava/util/Iterator;"))
-    private static void initFabricConfig(String modId, Class<?> config, CallbackInfo ci) {
-        if (modId.equals(FossilMod.MOD_ID)) {
+    @Shadow
+    protected String modid;
+
+    @Inject(method = "loadValuesFromJson", remap = false, at = @At(value = "INVOKE", target = "Ljava/util/Collection;forEach(Ljava/util/function/Consumer;)V"))
+    private void initFabricConfig(CallbackInfo ci) {
+        if (modid.equals(FossilMod.MOD_ID)) {
             Field[] allFields = FossilConfigImpl.class.getDeclaredFields();
             for (Field field : allFields) {
                 if (field.getAnnotation(MidnightConfig.Entry.class) != null) {