+++ forge/src/main/java/com/github/teamfossilsarcheology/fossil/forge/FossilMixinPluginImpl.java
@@ -2,6 +2,9 @@ package com.github.teamfossilsarcheology.fossil.forge;
 
 import net.minecraftforge.fml.loading.FMLLoader;
 
+/**
+ * @see com.github.teamfossilsarcheology.fossil.FossilMixinPlugin
+ */
 public class FossilMixinPluginImpl {
     public static boolean isModLoaded(String mod) {
         return FMLLoader.getLoadingModList().getModFileById(mod) != null;