+++ CHANGELOG.md
@@ -8,3 +8,6 @@
 - Deinonychus sleeping animation
 - The axolotl and tropical fish patterns when using the fish egg item
 - Dinopedia not being a bookshelf book
+
+### Mod compatibility
+- Fixed game not starting with Midnightlib version >= 1.9.0